+++ examples/27.ammDeposit.js
@@ -28,7 +28,14 @@ provider.engine.stop()
 
 const dvfConfig = {
   api: envVars.API_URL,
-  dataApi: envVars.DATA_API_URL
+  dataApi: envVars.DATA_API_URL,
+  useAuthHeader: true,
+  wallet: {
+    type: 'tradingKey',
+    meta: {
+      starkPrivateKey: starkPrivKey
+    }
+  }
   // Add more variables to override default values
 }
 
@@ -39,17 +46,14 @@ const dvfConfig = {
 
   const waitForDepositCreditedOnChain = require('./helpers/waitForDepositCreditedOnChain')
 
-  const token1 = 'ETH'
-  const token2 = 'USDT'
-  const depositETHResponse = await dvf.deposit(token1, 0.1, starkPrivKey)
-  const depositUSDTResponse = await dvf.deposit(token2, 1000, starkPrivKey)
-
-  if (process.env.WAIT_FOR_DEPOSIT_READY === 'true') {
+  if (process.env.DEPOSIT_FIRST === 'true') {
+    const depositETHResponse = await dvf.deposit('KON', 100, starkPrivKey)
+    const depositUSDTResponse = await dvf.deposit('DVF', 200, starkPrivKey)
     await waitForDepositCreditedOnChain(dvf, depositETHResponse)
     await waitForDepositCreditedOnChain(dvf, depositUSDTResponse)
   }
 
-  const pool = `${token1}${token2}`
+  const pool = 'KONDVF'
 
   // Amm deposit consist of 2 orders, one for each of the pool tokens.
   // The tokens need to be supplied in a specific ratio. This call fetches
@@ -57,17 +61,14 @@ const dvfConfig = {
   // amount for that token.
   const ammFundingOrderData = await dvf.getAmmFundingOrderData({
     pool,
-    token: 'ETH',
-    amount: 0.1
+    token: 'KON',
+    amount: 100
   })
-
-  // ammFundingOrderData can be inspected/validate if desired, before signing
-  // the orders it contains and posting them to Deversifi API.
 
   // This call signs the orders contained in the ammFundingOrderData before
   // posting them to Deversifi API. NOTE: if the orders are pre-signed, the
   // method will post them as is.
-  const ammPostFundingOrderResponse = await dvf.postAmmFundingOrder(
+  const ammPostFundingOrderResponse = await dvf.postAmmFundingOrders(
     ammFundingOrderData
   )