+++ src/lib/stark/ledger/createSignedOrder.js
@@ -1,11 +1,14 @@
 const Eth = require('@ledgerhq/hw-app-eth').default
-const { byContractAddress } = require('@ledgerhq/hw-app-eth/erc20')
 const DVFError = require('../../dvf/DVFError')
 const BN = require('bignumber.js')
 const _ = require('lodash')
 const selectTransport = require('../../ledger/selectTransport')
+const provideTokenSignature = require('../../ledger/provideTokenSignature')
 const getTokenAddressFromTokenInfoOrThrow = require('../../dvf/token/getTokenAddressFromTokenInfoOrThrow')
-const generateTestNetworkTokenData = require('../../ledger/generateTestNetworkTokenData')
+const swJS = require('starkware_crypto')
+const {
+  starkLimitOrderToMessageHash
+} = require('dvf-utils')
 
 const getPublicKey = async (eth, starkPath) => {
   const tempKey = (await eth.starkGetPublicKey(starkPath)).toString('hex')
@@ -15,7 +18,8 @@ const getPublicKey = async (eth, starkPath) => {
   }
 }
 
-module.exports = async (dvf, path, starkOrder, { returnStarkPublicKey = true } = {}) => {
+// Use signing on the starkMessage only as a fallback
+module.exports = async (dvf, path, starkOrder, { returnStarkPublicKey = true, starkMessage = null } = {}) => {
   const Transport = selectTransport(dvf.isBrowser)
 
   const buySymbol = _.findKey(dvf.config.tokenRegistry, {
@@ -31,119 +35,54 @@ module.exports = async (dvf, path, starkOrder, { returnStarkPublicKey = true } =
 
   const transport = await Transport.create()
   const eth = new Eth(transport)
-  const {address} = await eth.getAddress(path)
+  const { address } = await eth.getAddress(path)
   const starkPath = dvf.stark.ledger.getPath(address)
 
   const starkPublicKey = returnStarkPublicKey
     ? await getPublicKey(eth, starkPath)
     : null
 
+  let buyTokenAddress = null
+  let sellTokenAddress = null
   try {
-    // TODO Extract below code to a utility method
-    // to be used for both buy as sell tokens and
-    // for transfer method as well as well
-
-    let buyTokenAddress = getTokenAddressFromTokenInfoOrThrow(buyTokenInfo, 'ETHEREUM')
-    if (buyTokenAddress) {
-      const buyTokenInfo = byContractAddress(buyTokenAddress)
-      buyTokenAddress = buyTokenAddress.substr(2)
-      if (buyTokenInfo) {
-        await eth.provideERC20TokenInformation(buyTokenInfo)
-      } else {
-        if (buyTokenAddress === 'dddddd4301a082e62e84e43f474f044423921918') {
-          await eth.provideERC20TokenInformation({
-            data: Buffer.from(
-              `03445646DDdddd4301A082e62E84e43F474f04442392191800000012000000013045022100bd8a55c10b02bbe70f7266be7f5f5e7132140623b6de3fa27bdd820f11baa0d902207eb91acba7c2c5131d8285f9eba2f0d06bc9be3b4dfc29d05b0f25aa3b620a41`,
-              'hex'
-            )
-          })
-        } else if (buyTokenAddress === 'dddd0e38d30dd29c683033fa0132f868597763ab') {
-          await eth.provideERC20TokenInformation({
-            data: Buffer.from(
-              `0478445646dddd0e38d30dd29c683033fa0132f868597763ab00000012000000013045022100b6a0d84315a9935d0245336bb12ef30fc4b4592d30407f33ab695f01fc445eef02200a74e7ee316e753e127de5e05a3fefade56b025a7c84d107594808582399872d`,
-              'hex'
-            )
-          })
-        } else if (buyTokenAddress === '0a0e3bfd5a8ce610e735d4469bc1b3b130402267') {
-          await eth.provideERC20TokenInformation({
-            data: Buffer.from(
-              '034552500a0e3bfd5a8ce610e735d4469bc1b3b1304022670000001200000001304402204fc34364bdba254200371327e2f6cff6f6e07f488ee6d71f95499429be6e76fc02201d222d7c93a395bed86f9e0cf1b8b3f6fabec8535807d013783a9201fec30531',
-              'hex'
-            )
-          })
-        } else if (dvf.config.ethereumChainId !== 1) {
-          const tokenData = generateTestNetworkTokenData(buyTokenAddress, dvf.config.ethereumChainId)
-          await eth.provideERC20TokenInformation(tokenData)
-        } else {
-          throw new DVFError('LEDGER_TOKENINFO_ERR')
-        }
-      }
-    } else {
-      buyTokenAddress = null
+    buyTokenAddress = getTokenAddressFromTokenInfoOrThrow(buyTokenInfo, 'ETHEREUM')
+    sellTokenAddress = getTokenAddressFromTokenInfoOrThrow(sellTokenInfo, 'ETHEREUM')
+    const [buySignature, sellSignature] = await Promise.all([
+      provideTokenSignature(dvf, eth, buyTokenAddress),
+      provideTokenSignature(dvf, eth, sellTokenAddress)
+    ])
+    if (buySignature.unsafeSign || sellSignature.unsafeSign) {
+      const message = starkMessage || starkLimitOrderToMessageHash(swJS)(starkOrder)
+      const paddedMessage = `0x${message.padEnd(64, '0').substr(-64)}`
+      const starkSignature = await eth.starkUnsafeSign(
+        starkPath,
+        paddedMessage
+      )
+      await transport.close()
+      return { starkSignature, starkPublicKey }
     }
-
-    // TODO Extract below code to a utility method
-    // to be used for both buy as sell tokens and
-    // for transfer method as well as well
-
-    let sellTokenAddress = getTokenAddressFromTokenInfoOrThrow(sellTokenInfo, 'ETHEREUM')
-    if (sellTokenAddress) {
-      const sellTokenInfo = byContractAddress(sellTokenAddress)
-      sellTokenAddress = sellTokenAddress.substr(2)
-      if (sellTokenInfo) {
-        await eth.provideERC20TokenInformation(sellTokenInfo)
-      } else {
-        if (sellTokenAddress === 'dddddd4301a082e62e84e43f474f044423921918') {
-          await eth.provideERC20TokenInformation({
-            data: Buffer.from(
-              `03445646DDdddd4301A082e62E84e43F474f04442392191800000012000000013045022100bd8a55c10b02bbe70f7266be7f5f5e7132140623b6de3fa27bdd820f11baa0d902207eb91acba7c2c5131d8285f9eba2f0d06bc9be3b4dfc29d05b0f25aa3b620a41`,
-              'hex'
-            )
-          })
-        } else if (sellTokenAddress === 'dddd0e38d30dd29c683033fa0132f868597763ab') {
-          await eth.provideERC20TokenInformation({
-            data: Buffer.from(
-              `0478445646dddd0e38d30dd29c683033fa0132f868597763ab00000012000000013045022100b6a0d84315a9935d0245336bb12ef30fc4b4592d30407f33ab695f01fc445eef02200a74e7ee316e753e127de5e05a3fefade56b025a7c84d107594808582399872d`,
-              'hex'
-            )
-          })
-        } else if (sellTokenAddress === '0a0e3bfd5a8ce610e735d4469bc1b3b130402267') {
-          await eth.provideERC20TokenInformation({
-            data: Buffer.from(
-              '034552500a0e3bfd5a8ce610e735d4469bc1b3b1304022670000001200000001304402204fc34364bdba254200371327e2f6cff6f6e07f488ee6d71f95499429be6e76fc02201d222d7c93a395bed86f9e0cf1b8b3f6fabec8535807d013783a9201fec30531',
-              'hex'
-            )
-          })
-        } else if (dvf.config.ethereumChainId !== 1) {
-          const tokenData = generateTestNetworkTokenData(sellTokenAddress, dvf.config.ethereumChainId)
-          await eth.provideERC20TokenInformation(tokenData)
-        } else {
-          throw new DVFError('LEDGER_TOKENINFO_ERR')
-        }
-      }
-    } else {
-      sellTokenAddress = null
-    }
-
-    const starkSignature = await eth.starkSignOrder_v2(
-      starkPath,
-      sellTokenAddress,
-      sellSymbol === 'ETH' ? 'eth' : 'erc20',
-      new BN(sellTokenInfo.quantization),
-      null,
-      buyTokenAddress,
-      buySymbol === 'ETH' ? 'eth' : 'erc20',
-      new BN(buyTokenInfo.quantization),
-      null,
-      starkOrder.vaultIdSell,
-      starkOrder.vaultIdBuy,
-      new BN(starkOrder.amountSell),
-      new BN(starkOrder.amountBuy),
-      starkOrder.nonce,
-      starkOrder.expirationTimestamp
-    )
-    return { starkPublicKey, starkSignature }
-  } finally {
+  } catch (e) {
     await transport.close()
+    throw new DVFError('LEDGER_TOKENINFO_ERR')
   }
+
+  const starkSignature = await eth.starkSignOrder_v2(
+    starkPath,
+    sellTokenAddress,
+    sellSymbol === 'ETH' ? 'eth' : 'erc20',
+    new BN(sellTokenInfo.quantization),
+    null,
+    buyTokenAddress,
+    buySymbol === 'ETH' ? 'eth' : 'erc20',
+    new BN(buyTokenInfo.quantization),
+    null,
+    starkOrder.vaultIdSell,
+    starkOrder.vaultIdBuy,
+    new BN(starkOrder.amountSell),
+    new BN(starkOrder.amountBuy),
+    starkOrder.nonce,
+    starkOrder.expirationTimestamp
+  )
+  await transport.close()
+  return { starkPublicKey, starkSignature }
 }