+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.1.3 - 2021-11-15
+### Changed
+- Add getAmmFundingOrder method to track AMM operations
+
+## 3.1.2 - 2021-11-10
+### Changed
+- Add poolTvlHistory endpoint
+
+## 3.1.1 - 2021-11-10
+### Changed
+- When user denies register TX, respond with more friendly message
+
+## 3.1.0 - 2021-11-04
+### Changed
+- Assuming default web3 provider is an Ethereum mainnet or testnet
+- `useAuthHeader` authentication mode for all examples
+
 ## 3.0.3 - 2021-10-22
 ### Changed
 - Add manual xDVF support